requirements.txt
alembic==1.7.5
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.4.1
asttokens==2.0.5
backcall==0.2.0
black==21.12b0
cart==1.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
config==0.5.1
cryptography==36.0.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.1
Django==3.2.9
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-graphql-jwt==0.3.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
executing==0.8.2
filelock==3.0.12
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.2
idna==3.3
ipython==8.0.0
jedi==0.18.1
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
mysql-connector-python==8.0.27
mysqlclient==2.0.3
numpy==1.22.0
oauthlib==3.2.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
promise==2.3
prompt-toolkit==3.0.24
pure-eval==0.2.1
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.11.2
PyJWT==2.3.0
python-decouple==3.5
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
Rx==1.6.1
singledispatch==3.7.0
six==1.15.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
stack-data==0.1.3
stripe==2.66.0
text-unidecode==1.3
tomli==1.2.3
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.4.4
wcwidth==0.2.5
